+++ src/components/Account.tsx
@@ -1,81 +1,48 @@
-import { useCallback, useMemo, useState } from 'react';
-import { utils, providers } from 'ethers';
 import Blockies from 'react-blockies';
-import styled from 'styled-components';
-import { Box, Text, Notification } from 'grommet';
-import { getNetworkInfo } from '../config';
-import { centerEllipsis, copyToClipboard } from '../utils/strings';
-import { usePoller } from '../hooks/usePoller';
-import Logger from '../utils/logger';
-
-// Initialize logger
-const logger = Logger('Account');
+import { styled, useTheme } from '@mui/material/styles';
+import { Box, Paper, Typography } from '@mui/material';
+import { useMemo } from 'react';
+import { centerEllipsis } from '../utils/strings';
+import Iconify from '../components/Iconify';
 
 export interface AccountProps {
-  account?: string;
-  provider?: providers.JsonRpcProvider;
+  account: string;
+  open?: boolean;
 }
 
 const AccountIcon = styled(Blockies)`
   border-radius: 50%;
 `;
 
-const AccountHash = styled(Text)`
-  margin: 0 8px;
-  cursor: pointer;
-`;
-
-export const Account = ({ account, provider }: AccountProps) => {
-  const [balance, setBalance] = useState<string>('');
-  const [notification, setNotification] = useState<boolean>(false);
-
+export const Account = ({ account, open }: AccountProps) => {
+  const theme = useTheme();
   const shortAccount = useMemo(() => centerEllipsis(account || ''), [account]);
 
-  const getBalance = useCallback(async () => {
-    try {
-      if (provider && account) {
-        const network = await provider.getNetwork();
-        const chain = getNetworkInfo(network.chainId);
-        provider
-          .getBalance(account)
-          .then((balance) =>
-            setBalance(
-              `(${Number(utils.formatEther(balance)).toFixed(2)} ${chain.currency})`
-            )
-          )
-          // eslint-disable-next-line no-console
-          .catch(console.error);
-      } else {
-        setBalance('');
-      }
-    } catch (err) {
-      logger.error(err);
-      setBalance('');
-    }
-  }, [provider, account]);
-
-  usePoller(getBalance, !!provider, 2000, 'Account balance');
-
-  if (!provider || !account) {
-    return null;
-  }
-
   return (
-    <Box
-      direction="row"
-      align="center"
-      style={{ boxShadow: 'none' }}
-      onClick={() => {
-        copyToClipboard(account);
-        setNotification(true);
-        setTimeout(() => setNotification(false), 1500);
+    <Paper
+      variant="outlined"
+      sx={{
+        display: 'flex',
+        alignItems: 'center',
+        marginRight: { md: theme.spacing(5) },
+        padding: '3px 4px 2px 4px'
       }}
     >
-      <AccountIcon seed={account} size={7} scale={4} />
-      <AccountHash size="small">
-        {shortAccount}&nbsp;{balance}
-      </AccountHash>
-      {notification && <Notification toast title="Copied to clipboard" status="normal" />}
-    </Box>
+      <Box marginRight="5px">
+        <AccountIcon seed={account} size={7} scale={4} />
+      </Box>
+      <Box>
+        <Typography color="primary" variant="body2">
+          {shortAccount}
+        </Typography>
+      </Box>
+      {open !== undefined && (
+        <Iconify
+          color="inherit"
+          icon={`akar-icons:chevron-${open ? 'up' : 'down'}`}
+          marginLeft={theme.spacing(1)}
+        />
+      )}
+    </Paper>
   );
 };
