+++ src/components/AccountInfo.tsx
@@ -1,11 +1,9 @@
-import Blockies from 'react-blockies';
-import { styled, useTheme } from '@mui/material/styles';
+import { useTheme } from '@mui/material/styles';
 import {
   Box,
   Card,
   CardHeader,
   CardContent,
-  Paper,
   Popover,
   Typography,
   Button,
@@ -13,52 +11,11 @@ import {
 } from '@mui/material';
 import { LoadingButton } from '@mui/lab';
 import { useState, useMemo, useCallback, useRef, useEffect } from 'react';
-import { centerEllipsis, copyToClipboard } from '../utils/strings';
+import { copyToClipboard } from '../utils/strings';
 import Iconify from '../components/Iconify';
 import { useAccount, useNetwork, useDisconnect } from 'wagmi';
 import { getNetworkInfo } from '../config';
-
-export interface AccountProps {
-  account: string;
-  open?: boolean;
-}
-
-const AccountIcon = styled(Blockies)`
-  border-radius: 50%;
-`;
-
-export const Account = ({ account, open }: AccountProps) => {
-  const theme = useTheme();
-  const shortAccount = useMemo(() => centerEllipsis(account || ''), [account]);
-
-  return (
-    <Paper
-      variant="outlined"
-      sx={{
-        display: 'flex',
-        alignItems: 'center',
-        marginRight: { md: theme.spacing(5) },
-        padding: '3px 4px 2px 4px'
-      }}
-    >
-      <Box marginRight="5px">
-        <AccountIcon seed={account} size={7} scale={4} />
-      </Box>
-      <Box>
-        <Typography color="primary" variant="body2">
-          {shortAccount}
-        </Typography>
-      </Box>
-      {open !== undefined && (
-        <Iconify
-          color="inherit"
-          icon={`akar-icons:chevron-${open ? 'up' : 'down'}`}
-          marginLeft={theme.spacing(1)}
-        />
-      )}
-    </Paper>
-  );
-};
+import { Account } from './Account';
 
 export const AccountInfo = () => {
   const theme = useTheme();
